类型判断bug修改 袁达 2018.04.23

+++ service/src/main/java/com/zlwon/server/service/impl/ApplicationCaseServiceImpl.java
@@ -157,7 +157,7 @@ public class ApplicationCaseServiceImpl implements ApplicationCaseService {
 		if(list!=null &&  list.size()>0){
 			if(list.size()==1){
 				//判断id是否一样
-				if(applicationCaseDto.getId()!=list.get(0).getId()){
+				if(applicationCaseDto.getId()!=list.get(0).getId().intValue()){
 					throw new  CommonException(StatusCode.DATA_IS_EXIST);
 				}
 			}else {
